Created pharc.py in src directory to allow for the exe creation.  Modified as needed. Created setup.py - run this using py2exe to generate distributable package.
$ python setup.py py2exe

^ Will create a build and dist directory in src folder
build directory may be deleted
dist directory should be zipped up - this is the deliverable.

Currently, images do NOT display - and all functions dealing with images:
	Images don't display anywhere
	Images can't be exported
	Images can't be opened in external viewer
	etc...

This has something to do with the change in directory structure when the .exe
is created.  I wasn't able to track down where the issue was, and the few
fixes I attempted didn't work.